[MENFORCER-323] - Improve output of "Some plugins are missing valid versions" error
The old output could be:
```
Some plugins are missing valid versions:(LATEST RELEASE SNAPSHOT are not allowed )
org.apache.maven.plugins:maven-release-plugin. The version currently in use is 2.5.3
org.apache.maven.plugins:maven-resources-plugin. The version currently in use is 3.0.2
```

If that if branch wasn't taken, then there wouldn't even be a newline, which would make the output even uglier.